backend/requirements.txt
fastapi
redis
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-multipart
langchain-groq
langchain
sqlalchemy
psycopg2-binary
